Bug fixed: ImageMemo Initialized in right time

+++ src/components/Memo/MemoInput.js
@@ -53,11 +53,16 @@ class MemoInput extends Component {
         console.log("handleDelete Mode");
 
         const url = encodeURIComponent(this.props.url);
-        var db = firebase.firestore();
+        const ask = window.confirm("정말 삭제하시겠습니까?");
+        if(ask) {
+            var db = firebase.firestore();
 
-        db.collection("User").doc(this.props.uid).collection("Url").doc(url).collection("Memos").doc(this.props.id).delete();
+            db.collection("User").doc(this.props.uid).collection("Url").doc(url).collection("Memos").doc(this.props.id).delete();
 
-        document.getElementById(`stickyMemo_${this.props.id}`).style.visibility = 'hidden';
+            document.getElementById(`stickyMemo_${this.props.id}`).style.visibility = 'hidden';
+        }else{
+
+        }
     };
